dst: bug and test fixes (#874)
* dst: wait for spawned goroutines to finish on restart command

Otherwise, some goroutines from the previous store (like block persistence)
could race with the new store. This is an artifact of restarting the DB
in-process. The approach used is to simply scan the running goroutine stacks,
ignoring any goroutines that existed at the start of the test.

* wal: sleep for defaultTickTime if no progress is made

Previously, runtime.Gosched was called. However, this could be an issue in
deterministic simulation tests since if other goroutines are sleeping, the
yielding goroutine will schedule itself before advancing faketime. Sleeping is
a quick fix for avoiding getting into this infinite loop, although we should
probably come up with a long-term fix for these types of cases.

* lsm: fix nil pointer in merge

* dst: close most recent ColumnStore instead of first on defer

We were closing the incorrect column store at the end of the test (in some
cases double closing).

* db: correctly discard table block contained in snapshot if block was persisted

Previously, the code attempted to do this by resetting the active block index.
However, the ULID was left untouched. This would cause a previously persisted
block to lose data when the new empty block's rotation happened. This commit
updates the full block when a NewTableBlock entry is found and the table
exists.

* table: enhance rotation log messages

These log messges were useful when debugging DST failures.