table: fix behavior on new table block error
Table block creation can fail for a variety of reasons, but we were only
handling the happy case. This commit only overwrites the active block if the
new table block was successfully created, and logs a noop record to the WAL on
error to not block subsequent wal records logs.